feat: generate check name based on stacktrace
If there was a false positive in the scanner, a similar ticket will be
created if the partner pushes an update, since the artifact id is
included in the path. We remove the artifact id by starting the hash at
the src directory and include the stacktrace, so that the ticket will
remain the same.